--- types/sinon/ts3.1/tslint.json
@@ -1,7 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "ban-types": false,
-        "unified-signatures": false
-    }
-}
